format code with forge fmt

+++ packages/perennial-oracle/contracts/interfaces/IPythFactory.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.13;
 
-import { IKeeperFactory } from "./IKeeperFactory.sol";
+import {IKeeperFactory} from "./IKeeperFactory.sol";
 
 interface IPythFactory is IKeeperFactory {
     // sig: 0x22445848
@@ -10,5 +10,5 @@ interface IPythFactory is IKeeperFactory {
 
 /// @dev PythStaticFee interface, this is not exposed in the AbstractPyth contract
 interface IPythStaticFee {
-    function singleUpdateFeeInWei() external view returns (uint);
+    function singleUpdateFeeInWei() external view returns (uint256);
 }
